DEV: Initial Commit (#1)
This PR adds the core functionality of the plugin. 

# Overview
There is a gap in Discourse Core when it comes to tracking a user's session... especially for anon users.

This plugin adds a `page_visits` table to Discourse Core that then commits a user's 

- user_id (if user is authenticated)
- ip_address
- user_agent
- topic_id (if visiting a topic route)
- post_ids (if visiting a topic route, we record every post_id that was read, even if it has been read before)
- visit_time (stored in ms)

for each visit session. This is available for both anonymous and authenticated users. A session is ended, and data is written to the DB on (one of) two conditions: 

1. They change pages (or routes). 
2. They trigger a `visiblitychange` event. 

While changing routes is a simple pattern to follow and catch, `visiblitychange` is not. Following [documentation](https://www.igvita.com/2015/11/20/dont-lose-user-and-app-state-use-page-visibility/) referred to by MDN, there is no _catch all_ solution for all desktop / mobile browsers, for tracking when a browser session has ended. The best and most widely accepted solution is to place your analytic callbacks in a 

```js
if (document.visibilityState === "hidden")
```

block, which this PR introduces.